Move dodging function to MainCharacter
WIP. Instead of returning a Response::Transition::State(), calling the
state fn from the state machine, return a variant of the state enum.
Then we match the variant in the state machine and call the transition
from there.

Reason being, even if the state function has a public identifier, I
think the proc macro prevents it from being called outside of that
module.