add constants and change input handling
Added constants for signals, callables, and nodes. Rather than depending
on &str everywhere, define constants. Should prevent typing errors as
well as limit number of locations to update if a name changes.

Also change to platformer input handling. This is a work in progress as
jumping, climbing, falling/gravity still need to be added. Probably the
next PR.