wip add attacking
Implementing attacking behavior for enemy AI. Even with the small amount of signals, and by extension callables, that I'm currently working with I am finding
it unergonomic. Will commit recent changes and switch to master branch to have typed signals from gdext PR 1000.

It doesn't sound like much will change in relation to the current implementation of typed signals with the next gdext release, so may as well be
ahead of the curve.