refactor state machine
Refactors the player character state machine to prevent passing a godot
owned pointer around. The player character now just matches the state of
the state machine and calls the functions for itself.

While there is nothing inherently wrong with passing a Gd<>, it does add
an amount of cognitive overhead in remembering where and when binding
occurs on the pointer. With this PR, we don't have to worry about that
and the borrow checker will ensure borrowing rules are upheld.