add autoloader
The autoloader is responsible for connecting signals from the player and the player's UI (health, stamina, etc.).
The alternative would be for instances to hold references of the object. For example, the player object would need to obtain a pointer to the HealthBar located in the scene tree.
That seems fragile and error prone. With a little discipline I think a singleton-like node that just handles signals would prove more robust.

change name of singleton

remove unused constants

rename singleton in godot